+++ src/main/java/com/nikolay/springmimimimetr/repositories/ViewRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ViewRepository extends JpaRepository<View, Long> {
+public interface ViewRepository extends JpaRepository<View, Integer> {
 
     public View findOneByCandidateAndUsername(Candidate candidate, String username);
